+++ utils/mrisurf_timeStep.c
@@ -57,13 +57,13 @@ static int MRISAsynchronousTimeStep_optionalDxDyDzUpdate_svi(
   float x, float y, float z) 
 {
   const size_t numSubvolsPerEdge   = MRISAsynchronousTimeStep_optionalDxDyDzUpdate_numSubvolsPerEdge;
-  int const svxLo = MIN(numSubvolsPerEdge-1,(int)((x - ctx->xSubvolVerge - ctx->xLo)/ctx->xSubvolLen));
-  int const svyLo = MIN(numSubvolsPerEdge-1,(int)((y - ctx->ySubvolVerge - ctx->yLo)/ctx->ySubvolLen));
-  int const svzLo = MIN(numSubvolsPerEdge-1,(int)((z - ctx->zSubvolVerge - ctx->zLo)/ctx->zSubvolLen));
+  int const svxLo = MIN((int)numSubvolsPerEdge-1,(int)((x - ctx->xSubvolVerge - ctx->xLo)/ctx->xSubvolLen));
+  int const svyLo = MIN((int)numSubvolsPerEdge-1,(int)((y - ctx->ySubvolVerge - ctx->yLo)/ctx->ySubvolLen));
+  int const svzLo = MIN((int)numSubvolsPerEdge-1,(int)((z - ctx->zSubvolVerge - ctx->zLo)/ctx->zSubvolLen));
 
-  int const svxHi = MIN(numSubvolsPerEdge-1,(int)((x + ctx->xSubvolVerge - ctx->xLo)/ctx->xSubvolLen));
-  int const svyHi = MIN(numSubvolsPerEdge-1,(int)((y + ctx->ySubvolVerge - ctx->yLo)/ctx->ySubvolLen));
-  int const svzHi = MIN(numSubvolsPerEdge-1,(int)((z + ctx->zSubvolVerge - ctx->zLo)/ctx->zSubvolLen));
+  int const svxHi = MIN((int)numSubvolsPerEdge-1,(int)((x + ctx->xSubvolVerge - ctx->xLo)/ctx->xSubvolLen));
+  int const svyHi = MIN((int)numSubvolsPerEdge-1,(int)((y + ctx->ySubvolVerge - ctx->yLo)/ctx->ySubvolLen));
+  int const svzHi = MIN((int)numSubvolsPerEdge-1,(int)((z + ctx->zSubvolVerge - ctx->zLo)/ctx->zSubvolLen));
 
   if (svxLo != svxHi || svyLo != svyHi || svzLo != svzHi) 
     return MRISAsynchronousTimeStep_optionalDxDyDzUpdate_numSubvolsPerThread - 1;       // the svi for any items not deep inside a subvolume
@@ -625,7 +625,7 @@ static void mrisAsynchronousTimeStep_optionalDxDyDzUpdate( // BEVIN mris_make_su
       
       // Choose the subvolume to put this face into
       //
-      int const svi = (sviLo==sviHi) ? sviLo : numSubvolsPerThread-1;   // if not same subvol, into the shared subvol
+      int const svi = (sviLo==sviHi) ? sviLo : (int)numSubvolsPerThread-1;   // if not same subvol, into the shared subvol
 
       // Assign the face to the subvolume
       //
@@ -718,13 +718,13 @@ static void mrisAsynchronousTimeStep_optionalDxDyDzUpdate( // BEVIN mris_make_su
   { int * temp = (int*)malloc(sizeof(int)*mris->nvertices);
     
     int svi;
-    for (svi = 0; svi < numSubvolsPerThread; svi++) {
+    for (svi = 0; svi < (int)numSubvolsPerThread; svi++) {
     
       // build the list in the temp
       //
       int tempSize = 0;
       
-      int tid;
+      size_t tid;
       for (tid = 0; tid < numThreads; tid++) {
         SubvolInfo* subvol = subvols + tid*numSubvolsPerThread + svi;
         int vno = subvol->firstVnoPlus1 - 1;
